pyproject.toml
[project]
name = "ai-agent-gemma3"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dotenv>=0.9.9",
    "fastembed>=0.7.0",
    "huggingface-hub>=0.31.2",
    "llama-cpp-python>=0.3.8",
    "openai>=1.78.1",
    "pynvml>=12.0.0",
    "qdrant-client>=1.14.2",
    "streamlit>=1.45.1",
    "tabulate>=0.9.0",
    "torch>=2.7.0",
]
